+++ index.html
@@ -14,7 +14,7 @@
     <main>
       <div id="left-container">
         <h1>Garden of Verses</h1>
-        <form>
+        <form id="search-form">
           <fieldset id="input-field">
             <legend>Input fields</legend>
 
@@ -79,6 +79,6 @@
       </div>
     </main>
 
-    <script src="script.js"></script>
+    <script type="module" src="index.js"></script>
   </body>
 </html>
