+++ src/config.js
@@ -1,0 +1,1 @@
+export const POETRY_DB_BASE_URL = 'https://poetrydb.org'
